[PTYTHON] Migrate VTA TIR passes to the new pass manager.
